refactor(frontend): 去掉 any 并拆分超长通知页
Login/UserList/authSlice/storage 改为明确类型。通知列表与模板页拆成
列定义、详情抽屉与表单组件，单文件不超过 300 行且无行为变化。
