+++ frontend/src/pages/user/UserList.tsx
@@ -1,9 +1,13 @@
 import React, { useState, useEffect } from 'react';
-import { Table, Card, Button, Input, Select, Space, Modal, Form, message, Tag } from 'antd';
-import { PlusOutlined, SearchOutlined, EditOutlined, DeleteOutlined, ReloadOutlined } from '@ant-design/icons';
-import type { ColumnsType } from 'antd/es/table';
+import { Table, Card, Button, Input, Select, Space, Modal, Form, message } from 'antd';
+import { PlusOutlined, SearchOutlined, ReloadOutlined } from '@ant-design/icons';
 
 import { getUserList, UserListItem, createUser, updateUser, deleteUser } from '@/api/user';
+import { getUserColumns } from './userColumns';
+
+function isFormValidateError(error: unknown): boolean {
+  return typeof error === 'object' && error !== null && 'errorFields' in error;
+}
 
 const UserList: React.FC = () => {
   const [loading, setLoading] = useState(false);
@@ -98,88 +102,16 @@ const UserList: React.FC = () => {
       }
       setModalVisible(false);
       loadUsers();
-    } catch (error: any) {
-      if (error.errorFields) return; // 表单校验错误
+    } catch (error: unknown) {
+      if (isFormValidateError(error)) return;
       message.error(editingUser ? '更新失败' : '创建失败');
     }
   };
 
-  const columns: ColumnsType<UserListItem> = [
-    {
-      title: '用户名',
-      dataIndex: 'username',
-      key: 'username',
-      width: 120,
-    },
-    {
-      title: '真实姓名',
-      dataIndex: 'real_name',
-      key: 'real_name',
-      width: 100,
-    },
-    {
-      title: '邮箱',
-      dataIndex: 'email',
-      key: 'email',
-      width: 180,
-    },
-    {
-      title: '手机号',
-      dataIndex: 'phone',
-      key: 'phone',
-      width: 130,
-    },
-    {
-      title: '部门',
-      dataIndex: 'department_name',
-      key: 'department_name',
-      width: 100,
-    },
-    {
-      title: '职位',
-      dataIndex: 'position_name',
-      key: 'position_name',
-      width: 100,
-    },
-    {
-      title: '状态',
-      dataIndex: 'status',
-      key: 'status',
-      width: 80,
-      render: (status: number) => {
-        const statusMap: Record<number, { color: string; text: string }> = {
-          0: { color: 'success', text: '正常' },
-          1: { color: 'error', text: '禁用' },
-          2: { color: 'warning', text: '锁定' },
-        };
-        const info = statusMap[status] || statusMap[0];
-        return <Tag color={info.color}>{info.text}</Tag>;
-      },
-    },
-    {
-      title: '最后登录',
-      dataIndex: 'last_login_at',
-      key: 'last_login_at',
-      width: 160,
-      render: (time: string) => time || '-',
-    },
-    {
-      title: '操作',
-      key: 'action',
-      width: 150,
-      fixed: 'right',
-      render: (_, record) => (
-        <Space>
-          <Button type="link" size="small" icon={<EditOutlined />} onClick={() => handleEdit(record)}>
-            编辑
-          </Button>
-          <Button type="link" size="small" danger icon={<DeleteOutlined />} onClick={() => handleDelete(record)}>
-            删除
-          </Button>
-        </Space>
-      ),
-    },
-  ];
+  const columns = getUserColumns({
+    onEdit: handleEdit,
+    onDelete: handleDelete,
+  });
 
   return (
     <Card>
